Create first input of ubigeo section
Change the inputs to selects elements in the
ubication section. Add a new file with all the
data of this section in a json and use it in
the index file. Create 2 new states to this.
Map the data into a set and use the state
in the "department" select element as a prop